refactor(gui): extract header chrome builders into HeaderMixin
Move the three persistent-chrome builders (top navigation bar, status
banner, info bar) off WinpodxWindow into a private mixin so
main_window.py can focus on overall window orchestration:

- HeaderMixin (src/winpodx/gui/_main_window_header.py)
  _build_top_bar (logo + tabs + pod chip + start/stop), _build_status_banner
  (warning banner shown while pod is not running), _build_info_bar
  (status text + backend + resource summary).

Also moved _on_rdp_test into LogsMixin where it logically belongs —
it uses _log_append + log_signal and is invoked from the Logs/Tools
page.

WinpodxWindow now reads
`class WinpodxWindow(AppCrudMixin, HeaderMixin, InfoPageMixin, LogsMixin,
MaintenanceMixin, PodStatusMixin, QMainWindow)`.

Unicode-escape sequences (● etc.) in setText calls were folded
to their literal UTF-8 forms for readability; no behaviour change.
Tests: 1096 passed.

main_window.py: 1588 -> 1371 lines (-217). Cumulative: 2745 -> 1371 (-50%).
